fix another int overflow

+++ src/main/java/org/esa/beam/opendap/ui/DownloadAction.java
@@ -60,12 +60,12 @@ class DownloadAction implements ActionListener, DAPDownloader.FileCountProvider 
         final DAPDownloader downloader = new DAPDownloader(dapURIs, fileURIs, this, pm);
         final File targetDirectory = parameterProvider.getTargetDirectory();
         if (activeDownloaders.isEmpty()) {
-            pm.beginTask("", parameterProvider.getDatasize() / 1024);
+            pm.beginTask("", (int) (parameterProvider.getDatasize() / 1024.0));
             pm.worked(0);
             pm.resetStartTime();
             filesToDownloadCount = dapURIs.size() + fileURIs.size();
         } else {
-            pm.updateTask(parameterProvider.getDatasize() / 1024);
+            pm.updateTask((int) (parameterProvider.getDatasize() / 1024.0));
             filesToDownloadCount += dapURIs.size() + fileURIs.size();
         }
         DownloadWorker downloadWorker = new DownloadWorker(downloader, targetDirectory);
@@ -141,7 +141,7 @@ class DownloadAction implements ActionListener, DAPDownloader.FileCountProvider 
 
         void reset();
 
-        int getDatasize();
+        double getDatasize();
 
         File getTargetDirectory();
